refactor: rename Chrome extension package
