+++ extensions/geolibre-chrome/PRIVACY.md
