+++ extensions/geolibre-chrome/icons/geolibre-128.png
