+++ extensions/geolibre-chrome/icons/geolibre-32.png
