+++ extensions/geolibre-chrome/manifest.json
