+++ extensions/geolibre-chrome/popup.css
