+++ extensions/geolibre-chrome/scanner.mjs
