+++ scripts/package-chrome-extension.mjs
@@ -4,7 +4,7 @@ import { fileURLToPath } from "node:url";
 import { zipSync } from "fflate";
 
 const root = resolve(dirname(fileURLToPath(import.meta.url)), "..");
-const extensionDir = resolve(root, "extensions/geolibre-data-opener");
+const extensionDir = resolve(root, "extensions/geolibre-chrome");
 const manifest = JSON.parse(await readFile(resolve(extensionDir, "manifest.json"), "utf8"));
 const runtimeFiles = [
   "manifest.json",
@@ -24,7 +24,7 @@ for (const path of runtimeFiles)
   archive[path] = new Uint8Array(await readFile(resolve(extensionDir, path)));
 
 const outputDir = resolve(root, "dist");
-const output = resolve(outputDir, `geolibre-data-opener-${manifest.version}.zip`);
+const output = resolve(outputDir, `geolibre-chrome-${manifest.version}.zip`);
 await mkdir(outputDir, { recursive: true });
 await writeFile(output, zipSync(archive, { level: 9, mtime: new Date(1980, 0, 1) }));
 console.log(output);
